Refactor file tree
