+++ core/insert_entity.py
